Rimosso codice per arduino
